read_clip_tbl {clipr}	R Documentation

Transforms output of read_clip() into data frame.

Description

Transforms clipped content into a data frame by putting read_clip() output by using read.table().

Usage

read_clip_tbl(x = read_clip(), ...)

Arguments

x
Defaults to reading from the clipboard, but can be substituted by a character vector already generated by read_clip().

...
Options to pass to read.table(). The following read.table() arguments will be passed by default, but can be overridden by specifying them when calling read_clip_tbl:

header

TRUE

sep

"\t"

row.names

1

stringsAsFactors

FALSE

na.strings

c("NA", "")

strip.white

TRUE

Value

A data frame with the contents of the clipboard. If the system clipboard is empty, returns NULL

[Package clipr version 0.6.0 Index]