write_last_clip {clipr}	R Documentation

Write contents of the last R expression to the clipboard

Description

Write contents of the last R expression to the clipboard

Usage

write_last_clip(...)

Arguments

...
Pass other options to write_clip().

Note

This is a wrapper function for write_clip(.Last.value)

[Package clipr version 0.6.0 Index]