clipr_available {clipr}	R Documentation

Is the system clipboard available?

Description

Checks to see if the system clipboard is write-able/read-able. This may be useful if you are developing a package that relies on clipr and need to ensure that it will skip tests on machines (e.g. CRAN, Travis) where the system clipboard may not be available.

Usage

clipr_available()

dr_clipr()

Value

clipr_available returns a boolean value.

dr_clipr prints an informative message to the console with software and system configuration requirements if clipr is not available (invisibly returns the same string)

Examples

## Not run: 
# When using testthat:
library(testthat)
skip_if_not(clipr_available())

## End(Not run)

[Package clipr version 0.4.1 Index]