read_clip {clipr}	R Documentation

Read clipboard

Description

Read the contents of the system clipboard into a character vector.

Usage

read_clip()

Value

A character vector with the contents of the clipboard. If the system clipboard is empty, returns NULL

Note

read_clip() will not try to guess at how to parse copied text. If you are copying tabular data, it is suggested that you use read_clip_tbl().

Examples

## Not run: 
clip_text <- read_clip()

## End(Not run)

[Package clipr version 0.6.0 Index]